backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.97.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0

# Additional dependencies for enhanced backend
sqlalchemy==2.0.31
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.4.0
